Refactor: Move user sync primitives to UserSyncRepository
Moves `changeUserSecurity`, `saveKeyIv`, `replacedUrl`, `checkIfUserExists`,
`processUserAfterCreation`, `uploadNewUser`, and `updateExistingUser` from
`UploadToShelfService` into `UserSyncRepository` and its implementation
`UserRepositoryImpl`.

Updates `UploadToShelfService` to call these centralized user management logic
methods, maintaining it as the coordinator. Extracts `healthRepository` dependency
from `uploadNewUser` into a callback lambda to avoid circular dependencies.
